+++ lib/bootstrap.js
@@ -29,7 +29,7 @@ module.exports = function (config) {
     AquaJsPrecog = require('aquajs-precog');
     initPrecogLgger(app, config.app.logconfpath, AquaJsLogger.getWinston(), $logger);
   }
-  require('./express')(app);
+  require('./middleware')(app);
   if (config.app.loadModel && !$initModel && fs.existsSync($dirPaths.modelsDir)) {
     $enableWaterline=(config.connectionType==='mongo-orm');
     $enablePersist =(config.connectionType==='oracle-orm');
